// frontend/src/types/category.types.ts
export type NavCategory =
    "All"
    | "News"
    | "Research"
    | "Developers"
    | "Finance"
    | "Entertainment"
    | "Sports"
    | "Community"
    | "Discover";

export const NAV_CATEGORIES: NavCategory[] = [
    "News",
    "Research",
    "Developers",
    "Finance",
    "Entertainment",
    "Sports",
    "Community",
    "Discover",
];